- moved exceptions in a single file - singleton design patern core - strategy design patern core, ml, dl
