--- src/server/core/exceptions/uninitializedexception.py
@@ -1,11 +1,0 @@
-class UninitializedException(Exception):
-    """Exception raised by the core modules for incomplete/incorrect initialization.
-
-    Attributes:
-        message -- explanation of the error
-    """
-
-    def __init__(self, message="Class has not been initialized"):
-        self.message = message
-        super().__init__(self.message)
-
