+++ src/server/core/interfaces.py
@@ -1,0 +1,3 @@
+class AnalyzerInterface:
+    def analyze(self, preprocessed) -> [int]:
+        pass
